+++ src/app/(DefaultLayout)/(ProfileLayout)/reservation/components/LoadingState.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+function LoadingState() {
+  return (
+    <div className="flex justify-center items-center h-[20rem]">
+      <p className="text-14-m text-gray-500">불러오는 중...</p>
+    </div>
+  );
+}
+
+export default LoadingState;
